+++ main/assets/scripts/NodeClass.js
