+++ main/assets/scripts/drawGates.js
