+++ main/assets/scripts/getInfoPanel.js
