+++ main/assets/scripts/spatialHash.js
