+++ main/assets/scripts/start.js
