+++ main/assets/styles/level_select.css
@@ -16,7 +16,7 @@
   font-family: "Modak", serif;
   height: 100vh;
   width: 100%;
-  background-image: url(images/circuit.png);
+  background-image: url(../images/circuit.png);
   background-repeat: no-repeat;
   background-size: cover;
   background-position: center;
