+++ main/pages/index.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>GateKeeper - Start</title>
-    <link rel="stylesheet" href="../assets/start.css">
+    <link rel="stylesheet" href="../assets/styles/start.css">
     
     <!-- Link the google fonts -->
     <link rel="preconnect" href="https://fonts.googleapis.com">
@@ -537,8 +537,8 @@
 
     
     <script src="https://unpkg.com/aos@2.3.1/dist/aos.js"></script>
-    <script src="../assets/start.js"></script>
-    <script src="../assets/sound_effect.js"></script>
-    <script src="../assets/language.js"></script>
+    <script src="../assets/scripts/start.js"></script>
+    <script src="../assets/scripts/sound_effect.js"></script>
+    <script src="../assets/scripts/language.js"></script>
 </body>
 </html>
