+++ main/pages/play.html
@@ -3,7 +3,7 @@
   <head>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/1.9.4/p5.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/1.9.4/addons/p5.sound.min.js"></script>
-    <link rel="stylesheet" type="text/css" href="../assets/style.css">
+    <link rel="stylesheet" type="text/css" href="../assets/styles/style.css">
     <title>GateKeeper - Main</title>
     <meta charset="utf-8" />
 
@@ -106,16 +106,16 @@
 
 
     </main>
-    <script src="../assets/spatialHash.js"></script>
+    <script src="../assets/scripts/spatialHash.js"></script>
     
-    <script src="../assets/NodeClass.js"></script>
+    <script src="../assets/scripts/NodeClass.js"></script>
 
-    <script src="../assets/drawGates.js"></script>
+    <script src="../assets/scripts/drawGates.js"></script>
     
-    <script src="../assets/GameSounds.js"></script>
-    <script src="../assets/GateKeeperObjects.js"></script>
-    <script src="../assets/sound_effect.js"></script>
-    <script src="../assets/options.js"></script>
-    <script src="../assets/getInfoPanel.js"></script>
+    <script src="../assets/scripts/GameSounds.js"></script>
+    <script src="../assets/scripts/GateKeeperObjects.js"></script>
+    <script src="../assets/scripts/sound_effect.js"></script>
+    <script src="../assets/scripts/options.js"></script>
+    <script src="../assets/scripts/getInfoPanel.js"></script>
   </body>
 </html>
