+++ main/pages/policy.html
@@ -3,7 +3,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>GateKeeper - Terms & Privacy</title>
-    <link rel="stylesheet" href="/SoftwareEngGroupA/main/assets/policy.css">
+    <link rel="stylesheet" href="/SoftwareEngGroupA/main/assets/styles/policy.css">
 
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
@@ -102,6 +102,6 @@
 
 
 
-    <script src="policy.js"></script>
+    <script src="scripts/policy.js"></script>
 </body>
 </html>
