+++ main/pages/terms.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>GateKeeper - Terms & Privacy</title>
-    <link rel="stylesheet" href="/SoftwareEngGroupA/main/assets/terms.css">
+    <link rel="stylesheet" href="/SoftwareEngGroupA/main/assets/styles/terms.css">
 
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
@@ -78,6 +78,6 @@
 
 
 
-    <script src="terms.js"></script>
+    <script src="scripts/terms.js"></script>
 </body>
 </html>
